requirements.txt
alembic==1.9.2
attrs==21.2.0
autopep8==1.5.5
blinker==1.5
cachetools==5.2.0
certifi==2022.9.14
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
cryptography==39.0.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
google-auth==2.11.0
google-auth-oauthlib==0.5.3
gunicorn==20.1.0
idna==2.10
importlib-metadata==4.12.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.1.4
MarkupSafe==2.1.1
oauthlib==3.2.1
packaging==20.9
pluggy==0.13.1
psycopg2-binary==2.9.5
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.21
PyJWT==2.6.0
pyparsing==2.4.7
pytest==7.1.1
pytest-flask==1.2.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.1
toml==0.10.2
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.8.1
